refactor(win32): split input test module into a directory under line budget
